requirements.txt
anaconda-client==1.11.2
anaconda-navigator==2.4.0
anaconda-project==0.11.1
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.2
bleach==4.1.0
boltons==23.0.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
clyent==1.2.2
colorama==0.4.6
comm==0.2.1
conda==23.9.0
conda-build==3.28.4
conda-content-trust==0.1.3
conda_index==0.4.0
conda-pack==0.6.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
conda-repo-cli==1.0.88
conda-token==0.4.0
conda-verify==3.4.2
contourpy==1.2.0
cryptography==42.0.5
cycler==0.11.0
Cython==3.0.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
fire==0.5.0
fonttools==4.25.0
fsspec==2024.3.1
future==0.18.3
gmpy2==2.1.2
idna==3.4
imagecodecs==2023.1.23
imageio==2.33.1
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.3
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
kiwisolver==1.4.4
lazy_loader==0.3
libarchive-c==2.9
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
medmnist==2.2.3
menuinst==1.4.19
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
munkres==1.1.4
navigator-updater==0.4.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.0.8
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pillow==10.2.0
pip==22.3.1
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
psutil==5.9.0
pure-eval==0.2.2
pycosat==0.6.6
pycparser==2.21
pydicom==2.4.4
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==24.0.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-msssim==1.0.0
pytz==2023.3.post1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
scikit-image==0.22.0
scipy==1.12.0
Send2Trash==1.8.2
setuptools==65.6.3
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
termcolor==2.4.0
terminado==0.17.1
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.1+cu121
torchsummary==1.5.1
torchvision==0.16.1+cu121
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
truststore==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.4.0
urllib3==2.1.0
vboxapi==1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
wincertstore==0.2
zstandard==0.19.0
